requirements.txt
# Eve Framework - Visual Windows Automation
# Core dependencies

pyautogui>=0.9.54      # Mouse/keyboard control
opencv-python>=4.8.0   # Image template matching
pillow>=10.0.0         # Image processing
mss>=9.0.0             # Fast screenshots
pygetwindow>=0.0.9     # Window management
pywin32>=306           # Windows API access
numpy>=1.24.0          # Required by OpenCV
psutil>=5.9.0          # Process management for Sanderling

# Telegram notifications
python-telegram-bot>=20.0  # Telegram bot API
python-dotenv>=1.0.0       # Load environment variables from .env
